Move renderDataCell
